Final Version
separated content and data from each other and added active colors to the nav-links
