+++ src/components/Content.css
@@ -265,7 +265,7 @@
         flex-direction: column-reverse;
         position: relative;
         top: 0;
-        margin-top: 0;
+        margin-top: 100px;
     }
 
     .overview-istructure-sgeology{
@@ -354,7 +354,6 @@
 
     .content-img img{
         width: 60%;
-        margin-top: -100px;
     }
 
     .geology-img img{
